% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MATLABfuncs.R
\name{run_tSpace}
\alias{run_tSpace}
\title{Run the matlab version of the tSpace algorithm}
\usage{
run_tSpace(
  data,
  trajectories = 100,
  k = c(20, 30, 50, 100),
  l = "auto",
  metric = "euclidean",
  graphs = 5,
  landmarks = 20,
  voting_scheme = "exponential",
  label = "",
  matlab_version = "auto",
  matlab_path = "~/Documents/MATLAB"
)
}
\arguments{
\item{data}{A matrix-like object where cells correspond to rows and features correspond to columns}

\item{trajectories}{The number of trajectories to calculate.}

\item{k}{Number of nearest neighbors (passed to the k argument of wanderlust). If a vector of integers is provided (recommended), the tSpace calculation will be attempted on the smallest value in the vector and then reattempted on successively higher values of k should the algorithm fail. tSpace caclualtions that require an excessively high k value suggest the presence of outlier data points that should be removed from the input data for better performance.}

\item{l}{Number of nearest neighbors (passed to the l argumet of wanderlust). If set to 'auto' (default), this parameter will be 2/3 of k (rounded to the nearest integer). If k is a vector, l must be a vector the same length and all corresponding values must be less than k.}

\item{metric}{A distance metric accepted by the dist function in matlab (wanderlust): euclidean, correlation, cityblock, chebyshev, minkowski, hamming, mahalanobis, jaccard, cosine or spearman.}

\item{graphs}{Number of graphs over which to average trajectories (wanderlust).}

\item{landmarks}{Number of landmarks (wanderlust).}

\item{voting_scheme}{Arugment passed to wanderlust.}

\item{label}{A character string to be incorporated into the tSpace column labels. Default is an empty label.}

\item{matlab_version}{A character string specifying the active matlab version (e.g. "R2019a"). If set to "auto," the latest version of matlab will be used. Only works for mac users.}

\item{matlab_path}{The path to the MATLAB folder containing cyt and tSpace scripts.}
}
\value{
A \code{list} that includes the following elements:
\describe{
  \item{tPCs}{A cell by tPC \code{matrix} of the PCA reduction of the trajectory data.}
  \item{traj}{A cell by trajectory \code{matrix} containing the trajectories.}
  \item{clusters}{A \code{vector} containing the cluster definitions used to initiate tSpace.}
  \item{pPCA}{A cell by pPCA \code{matrix} of the PCA reduction of the parameters.}
}
}
\description{
This is an interface to the matlab implementation of the tSpace algorithm described in 
Dermadi et al, modified for improved error handling.
}
\author{
Kevin Brulois
}
