% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MATLABfuncs.R
\name{runWanderlust}
\alias{runWanderlust}
\title{Run the wanderlust algorithm}
\usage{
runWanderlust(
  data,
  s,
  k = c(20, 30, 50, 100),
  l = "auto",
  metric = "euclidean",
  graphs = 20,
  landmarks = 20,
  voting_scheme = "exponential",
  matlab_version = "auto",
  matlab_path = "~/Documents/MATLAB"
)
}
\arguments{
\item{data}{A matrix-like object where cells correspond to rows and features correspond to columns.}

\item{k}{Number of nearest neighbors. If a vector of integers is provided (recommended), the wanderlust calculation will be attempted on the smallest value in the vector and then reattempted on successively higher values of k should the algorithm fail. Wanderlust caclualtions that require an excessively high k value suggest the presence of outlier data points that should be removed from the input data for better performance.}

\item{l}{Number of nearest neighbors. If set to 'auto' (default), l will be 2/3 of k (rounded to the nearest integer). If k is a vector, l must be a vector the same length and all corresponding values must be less than k.}

\item{metric}{A distance metric accepted by the dist function in matlab: euclidean, correlation, cityblock, chebyshev, minkowski, hamming, mahalanobis, jaccard, cosine or spearman.}

\item{graphs}{Number of graphs over which to average trajectories.}

\item{landmarks}{Number of landmarks.}

\item{voting_scheme}{Arugment passed to wanderlust.}

\item{matlab_version}{A character string specifying the active matlab version (e.g. "R2019a"). If set to "auto," the latest version of matlab will be used. Only works for mac users.}

\item{matlab_path}{The path to the MATLAB folder containing cyt and tSpace scripts.}
}
\value{
A \code{data.frame} containing the cell labels and the wanderlust trajectory.
}
\description{
This is an interface to the wanderlust algorithm described in Bendall et al.
}
\author{
Kevin Brulois
}
